requirements 2.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
websockets==13.0

# Screen capture and vision
mss==9.0.1
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# Game control
pyautogui==0.9.54
pywinauto==0.6.8

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25

# HTTP client for LLM
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Optional: LLM (uncomment based on your choice)
# openai==1.10.0
# anthropic==0.18.0
# ollama==0.1.6

# Optional: YOLO (uncomment when using actual YOLO model)
# ultralytics==8.1.0
# torch==2.1.2
# torchvision==0.16.2
# torchaudio==2.1.2
